requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
requests==2.32.5
python-dotenv==1.2.1
gTTS==2.5.4
pyttsx3==2.99
pywin32==311
ollama==0.1.0
